+++ packages/quantic/coveo.d.ts
@@ -2,10 +2,25 @@
 import * as HeadlessTypes from './force-app/main/default/staticresources/coveoheadless/index';
 export * from './force-app/main/default/staticresources/coveoheadless/index';
 import * as AtomicTypes from './force-app/main/default/staticresources/atomicutils/types/utils-exports';
+import {LightningElement} from 'lwc';
+import {HeadlessEngine} from '../headless/dist/index';
 export * from './force-app/main/default/staticresources/atomicutils/types/utils-exports';
 declare global {
   // eslint-disable-next-line no-undef
   const CoveoHeadless: typeof HeadlessTypes;
   // eslint-disable-next-line no-undef
   const CoveoAtomicUtils: typeof AtomicTypes;
+
+  interface Window {
+    coveoHeadless: {
+      [engineId: string]: {
+        components: {
+          element: LightningElement;
+          initialized: boolean;
+        }[];
+        config: HeadlessConfigurationOptions;
+        engine: HeadlessEngine;
+      };
+    };
+  }
 }